Refactor configuration loading and telemetry formatting in OverlayApp. Introduce functions to sanitize loaded configurations and streamline telemetry data processing. Enhance lap logging and session type handling for improved accuracy and display. Update related tests to ensure functionality.

+++ overlay-rs/crates/gridglance-overlay/src/ipc/mod.rs
@@ -274,7 +274,7 @@ fn dispatch(state: &StateHandle, req: Request) -> Response {
                     &st.map.merge_pts_2,
                     st.map.pit_speed_ms as f32,
                     st.map.pit_lane_speed_pct as f32,
-                    st.map.cached_pit2.lane_speed_pct as f32,
+                    st.map.cached_pit2.lane_speed_pct,
                     true,
                 )
             };
